Renamed Tracker to Workload_Processor
