+++ osbenchmark/benchmark.py
@@ -38,7 +38,7 @@ from osbenchmark import version, actor, config, paths, \
     test_execution_orchestrator, results_publisher, \
         metrics, workload, chart_generator, exceptions, log
 from osbenchmark.builder import provision_config, builder
-from osbenchmark.tracker import tracker
+from osbenchmark.workload_processor import workload_processor
 from osbenchmark.utils import io, convert, process, console, net, opts, versions
 
 
@@ -902,7 +902,7 @@ def dispatch_sub_command(arg_parser, args, cfg):
             cfg.add(config.Scope.applicationOverride, "workload", "workload.name", args.workload)
             configure_connection_params(arg_parser, args, cfg)
 
-            tracker.create_workload(cfg)
+            workload_processor.create_workload(cfg)
         elif sub_command == "info":
             configure_workload_params(arg_parser, args, cfg)
             workload.workload_info(cfg)
