+++ tests/workload_processor/corpus_test.py
@@ -25,7 +25,7 @@ import json
 from unittest import mock
 from unittest.mock import call
 
-from osbenchmark.tracker import corpus
+from osbenchmark.workload_processor import corpus
 
 
 def serialize_doc(doc):
